Add hello_world.test
Add hello_world.test, with a simple test to check if the current user is
root or not. This test is expected to succeed, assuming that the script is
executed by a regular user, only if the script is reexecuted in a new
environment, for example one created with an UML kernel.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>